+++ src/pages/workflows/model/useWorkflowListActions.ts
@@ -2,11 +2,8 @@ import { useState } from "react";
 import { useNavigate } from "react-router";
 
 import { useCreateWorkflowShortcut } from "@/features/create-workflow";
-import {
-  type WorkflowResponse,
-  buildPath,
-  useToggleWorkflowActiveMutation,
-} from "@/shared";
+import { buildPath, useToggleWorkflowActiveMutation } from "@/shared";
+import type { WorkflowResponse } from "@/shared/api";
 
 export const useWorkflowListActions = () => {
   const navigate = useNavigate();
