+++ src/pages/workflows/ui/WorkflowRow.tsx
@@ -17,7 +17,7 @@ import {
   VStack,
 } from "@chakra-ui/react";
 
-import { type WorkflowResponse } from "@/shared";
+import type { WorkflowResponse } from "@/shared/api";
 
 import {
   getBuildProgressLabel,
