+++ src/shared/api/auth/exchange-auth.api.ts
@@ -1,0 +1,18 @@
+import type { ApiResponse } from "../../types";
+import { processApiResponse } from "../../utils";
+import { apiClient } from "../client";
+
+import type { LoginResponse } from "./types";
+
+export const exchangeAuthAPI = async (
+  exchangeCode: string,
+): Promise<LoginResponse> => {
+  const { data } = await apiClient.post<ApiResponse<LoginResponse>>(
+    "/auth/exchange",
+    {
+      exchangeCode,
+    },
+  );
+
+  return processApiResponse(data);
+};
