+++ src/shared/api/execution/execute-workflow.api.ts
@@ -1,0 +1,13 @@
+import type { ApiResponse } from "../../types";
+import { processApiResponse } from "../../utils";
+import { apiClient } from "../client";
+
+export const executeWorkflowAPI = async (
+  workflowId: string,
+): Promise<string> => {
+  const { data } = await apiClient.post<ApiResponse<string>>(
+    `/workflows/${workflowId}/execute`,
+  );
+
+  return processApiResponse(data);
+};
