+++ src/shared/api/execution/get-execution-list.api.ts
@@ -1,0 +1,15 @@
+import type { ApiResponse } from "../../types";
+import { processApiResponse } from "../../utils";
+import { apiClient } from "../client";
+
+import type { ExecutionDetail } from "./types";
+
+export const getExecutionListAPI = async (
+  workflowId: string,
+): Promise<ExecutionDetail[]> => {
+  const { data } = await apiClient.get<ApiResponse<ExecutionDetail[]>>(
+    `/workflows/${workflowId}/executions`,
+  );
+
+  return processApiResponse(data);
+};
