+++ src/shared/api/execution/index.ts
@@ -1,0 +1,13 @@
+import { executeWorkflowAPI } from "./execute-workflow.api";
+import { getExecutionListAPI } from "./get-execution-list.api";
+import { getExecutionAPI } from "./get-execution.api";
+import { rollbackExecutionAPI } from "./rollback-execution.api";
+
+export * from "./types";
+
+export const executionApi = {
+  execute: executeWorkflowAPI,
+  getList: getExecutionListAPI,
+  getById: getExecutionAPI,
+  rollback: rollbackExecutionAPI,
+};
