+++ src/shared/api/index.ts
@@ -1,6 +1,6 @@
-export * from "./auth.api";
 export * from "./client";
-export * from "./execution.api";
-export * from "./oauth-token.api";
-export * from "./template.api";
-export * from "./workflow.api";
+export * from "./auth";
+export * from "./execution";
+export * from "./oauth-token";
+export * from "./template";
+export * from "./workflow";
