+++ src/shared/api/oauth-token/types.ts
@@ -1,0 +1,10 @@
+export interface OAuthTokenSummary {
+  service: string;
+  connected: boolean;
+  accountEmail: string | null;
+  expiresAt: string | null;
+}
+
+export interface OAuthConnectResponse {
+  authUrl: string;
+}
